vue_backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
Django==3.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-redis==4.11.0
django-redis-cache==2.1.0
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.1
pycodestyle==2.5.0
pylint==2.4.4
PyMySQL==0.9.3
pytz==2019.3
qiniu==7.2.6
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
